pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xmlplaylist"
version = "0.1.0"
description = "Export JSON dat tracku do mAirList XML playlistu (.mlp)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["mairlist", "playlist", "xml", "radio", "mlp"]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
xmlplaylist = "xmlplaylist.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["xmlplaylist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["xmlplaylist"]
omit = ["tests/*"]
